requirements.txt
# Core OCR Libraries
easyocr>=1.7.0
pytesseract>=0.3.10
paddleocr>=2.7.0.3

# Document Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.3
python-docx>=1.1.0
Pillow>=10.0.0

# Language Detection
fast-langdetect>=0.2.0

# Table Extraction
camelot-py[cv]>=0.11.0
pdfplumber>=0.10.3
tabulate>=0.9.0

# Performance & Monitoring
psutil>=5.9.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.1
beautifulsoup4>=4.12.0
fpdf2>=2.7.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0

# Markdown
markdown>=3.5.0
markdownify>=0.11.6